Fixed OpenGL uniform buffer layout

+++ source/opengl/gl_fx_compiler.hpp
@@ -73,6 +73,6 @@ namespace reshade
 		std::stringstream _global_code, _global_uniforms;
 		const reshadefx::nodes::function_declaration_node *_current_function;
 		std::unordered_map<const reshadefx::nodes::function_declaration_node *, function> _functions;
-		size_t _uniform_storage_offset = 0, _uniform_buffer_size = 0;
+		int _uniform_storage_offset = 0, _uniform_buffer_size = 0;
 	};
 }
